requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.12.13
attrs==22.1.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.10.0
celery==5.2.6
certifi==2022.9.24
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dill==0.3.6
Django==4.1.3
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-environ==0.8.1
django-filter==22.1
django-stubs==1.13.0
django-stubs-ext==0.7.0
django-timezone-field==5.0
djangorestframework==3.14.0
exceptiongroup==1.0.4
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.1.0
idna==2.10
importlib-metadata==5.1.0
iniconfig==1.1.1
isort==5.10.1
kombu==5.2.4
lazy-object-proxy==1.8.0
lxml==4.8.0
Markdown==3.4.1
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
prompt-toolkit==3.0.33
psycopg2==2.9.3
psycopg2-binary==2.9.5
PyJWT==2.6.0
pylint==2.15.7
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
pytest==7.2.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.6
redis==3.4.1
requests==2.23.0
sentry-sdk==1.11.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
types-pytz==2022.6.0.1
types-PyYAML==6.0.12.2
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.11.0
zope.event==4.5.0
zope.interface==5.5.2
